Rename scattering/maht.h to scattering/maths.h.
