+++ src/scattering/maths.h
