+++ src/scattering/single_scattering_data_impl.h
@@ -11,7 +11,7 @@
  */
 #pragma once
 
-#include <scattering/math.h>
+#include <scattering/maths.h>
 #include <scattering/interpolation.h>
 #include <scattering/scattering_data_field.h>
 #include <scattering/sht.h>
