+++ src/scattering/test_math.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include <scattering/math.h>
+#include <scattering/maths.h>
 
 namespace scattering {
